move cm.go for go install compatiblity
