+++ version/version.go
@@ -1,6 +1,6 @@
 // Copyright Contributors to the Open Cluster Management project
 
-package cmcli
+package version
 
 import (
 	_ "embed"
